REFACTOR: move the files in the handlers package to the server package
