+++ internal/server/errors.go
@@ -1,4 +1,4 @@
-package handlers
+package server
 
 import (
 	"encoding/json"
