+++ internal/server/server.go
@@ -4,7 +4,6 @@ import (
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 
-	"github.com/alcb1310/bca-json/internal/handlers"
 	"github.com/alcb1310/bca-json/internal/utils"
 )
 
@@ -21,8 +20,8 @@ func NewServer() *Server {
 func (s *Server) MountHandlers() {
 	s.Router.Use(middleware.Logger)
 
-	s.Router.Get("/foo", utils.Make(handlers.HandleFoo))
+	s.Router.Get("/foo", utils.Make(s.HandleFoo))
 
-	s.Router.NotFound(utils.Make(handlers.NotFound))
-	s.Router.MethodNotAllowed(utils.Make(handlers.MethodNotAllowed))
+	s.Router.NotFound(utils.Make(NotFound))
+	s.Router.MethodNotAllowed(utils.Make(MethodNotAllowed))
 }
